chore(reorg): Move creator leapday modal to corresponding Svelte folder
